requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
altair==5.0.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
automium==0.2.6
automium-web==0.1.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bootstrap4==0.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cryptography==42.0.8
cssselect2==0.7.0
decorator==5.1.1
distlib==0.3.7
djando==0.2
Django==4.2.3
django-allauth==0.63.3
django-bootstrap4==23.2
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-payments==2.0.0
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
filelock==3.12.2
Flask==3.0.0
frozenlist==1.4.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
github==1.2.7
GitPython==3.1.32
gTTS==2.3.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
legacy==0.1.6
listview==0.1.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
multipledispatch==1.0.0
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.0
numpy==1.24.4
opencv-python==4.8.1.78
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
pbr==5.11.1
phonenumberslite==8.13.35
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
protobuf==4.23.4
psycopg2-binary==2.9.9
pyarrow==12.0.1
pycparser==2.22
pydeck==0.8.0
Pygments==2.15.1
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
Pympler==1.0.1
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.0
reportlab==4.2.2
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
soupsieve==2.5
sqlparse==0.4.4
stevedore==5.1.0
streamlit==1.25.0
svglib==1.5.1
tenacity==8.2.2
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
Trac==1.6
TracMovieMacro==0.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritools==4.0.3
urllib3==2.0.4
validators==0.20.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.0
xhtml2pdf==0.2.16
yarl==1.9.2
zipp==3.16.2
